review-feedback: WinLottoNumbersTest로 테스트 이동
