+++ docs/Lotto.md
@@ -50,4 +50,4 @@
 
 # 4단계 리뷰 반영
 - [X] LottoGame.takeBudget 메서드 삭제 (inputview와 중복)
-- [ ] WinLottoNumbersTest로 테스트 이동
+- [X] WinLottoNumbersTest로 테스트 이동
